Split storage implementations to sepearate files
